requirements.txt
gunicorn==19.0.0
django==2.2.5
djangorestframework==3.10.3
django-filter==2.2.0
coreapi==2.3.3
psycopg2==2.8.3
pandas==0.25.1